+++ CHANGELOG.md
@@ -34,7 +34,17 @@ All notable changes to this project will be documented in this file.
 
 ## [0.1.10 - dev] - 2023-12-29
 
+### Fixed
+
+-   Typos during command load/reload in specific guild (REST).
+
+### Changed
+
+-   Use `process.emitWarning()` for warnings instead of regular console logs.
+-   Throw an error if a global command registration/deletion fails instead of just logging (for legacy command registation).
+
 ### Removed
 
 -   `guildOnly` from docs examples. Closes [#42](https://github.com/underctrl-io/commandkit/issues/42)
 -   `guildOnly` deprecation warning.
+-   Emojis from logs, warnings, and errors.